feat: emit `TypedDocumentNode` alias in generated code
This change adds an exported type alias from each generated module based
on the `TypedDocumentNode` type from
`@graphql-typed-document-node/core`. Apollo Client and Villus are able
to consume this type and propagate result data and variables types
correctly. It looks like support for this type is also [coming][1] to
Apollo Angular and to GraphQL.js itself. For example,

    import { gql, useQuery } from '@apollo/client';

    const query = gql`
      query GitHubQuery($first: Int!) {
        viewer {
          repositories(first: $first) {
            nodes {
              id
              description
            }
          }
        }
      }
    ` as import("./__generated__/git-hub-query.ts").GithubQueryDocument;
    //   ^
    //   type assertion is co-located with query

    export default () => {
      const { data } = useQuery(query, { variables: { first: 100 } });
      //      ^                                     ^
      //      inferred type is `GithubQuery`        |
      //                                            |
      //                                 inferred type is `GithubQueryVariables`

      /* ... */
    };

As a result types can be hooked up with one type assertion that is
applied directly to the query expression.

[1]: https://github.com/dotansimha/graphql-typed-document-node#upcoming-built-in-support